Add files via upload
1.	Write a function to remove duplicate values in a string.
Ex: ‘Hello’ as input, the output should be ‘Helo’. ‘Good’ as input, the output should be ‘God’.

2.	Write a class and method under it to perform squareRoot, cuberoot and Nth root of a number.
Ex:- If I pass number 8 and 3, it should return cube root of 8. If I pass 16 and 4, it should retutn fourth root of 16.

3.	Write a function which takes a list as input and returns list in a reverse order.
Ex: [1,3,4,5,6] the output should be [6,5,4,3,1]

4.	Write a class which takes three inputs and returns the smallest and the largest number.
Ex: ClassName.smallest(1,2,3) should return 1. ClassName.largest(1,2,3) should return 3.